scripts/analysis/requirements.txt
bertopic==0.16.4
sentence-transformers==3.4.1
mecab-python3==1.0.9
unidic-lite==1.0.8
umap-learn==0.5.7
hdbscan==0.8.40
scikit-learn==1.6.1
matplotlib==3.10.0
plotly==5.24.1
kaleido==0.2.1
numpy==1.26.4
